+++ CHANGELOG.md
@@ -26,6 +26,8 @@
   error while connecting as a connect error.
 * Remove the need to pass `onDisconnection` handler to the `Transport.disconnect(...)` method and directly invoke
   callback from `Transport`.
+* Decouple the `ChangeSetParser` class from the javaee infrastructure and remove code used for testing
+  (i.e. `EntityChangeDataImpl`) from the main source tree into the test source tree.
 
 ### [v6.37](https://github.com/replicant4j/replicant/tree/v6.37) (2019-03-18)
 [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.36...v6.37)